+++ ignite/templates/app/options.go
@@ -10,6 +10,7 @@ type Options struct {
 	AddressPrefix    string
 	// IncludePrefixes is used to filter the files to include from the generator
 	IncludePrefixes []string
+	IsChainMinimal  bool
 }
 
 // Validate that options are usable.