+++ changelog.md
@@ -5,6 +5,7 @@
 ### Features
 
 - [#3839](https://github.com/ignite/cli/pull/3839) New structure for app scaffolding
+- [#3835](https://github.com/ignite/cli/pull/3835) Add `--minimal` flag to `scaffold chain` to scaffold a chain with the least amount of sdk modules
 
 ## [`v28.1.1`](https://github.com/ignite/cli/releases/tag/v28.1.1)